requirements.txt
eventlet==0.14.0
gunicorn==18.0
Flask==0.10.1
Flask-Migrate==1.2.0
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Flask-Negotiate==0.1.0
alembic==0.6.5
psycopg2==2.5.3
